requirements.txt
yfinance==0.1.55
ipykernel==5.3.4
scikit-learn==0.23.1
papermill==2.1.2
apache-airflow-providers-papermill==1.0.0
gitdb==4.0.5
gitpython==3.1.7
pytest==6.0.1